Rename TaskGroupCard to TaskBundleCard
