--- src/components/TaskGroupCard/index.js
@@ -1,1 +1,0 @@
-export { default } from './TaskGroupCard.vue';
